requirements.txt
tabulate
dotenv
openai
langchain
langchain_experimental
panda
streamlit
matplotlib
uvicorn
fastapi